refactor(drizzle): split postgresRaw into its own module
