+++ packages/alchemy/src/Drizzle/index.ts
@@ -1,3 +1,4 @@
 export * from "./Postgres.ts";
+export * from "./PostgresRaw.ts";
 export * from "./Providers.ts";
 export * from "./Schema.ts";
